requirements.txt
absl-py==2.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
dm-env==1.6
dm-tree==0.1.8
dm_control==1.0.25
etils==1.11.0
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
glfw==2.8.0
grpcio==1.68.0
gymnasium==1.0.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
labmaze==1.0.6
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
mujoco==3.2.5
networkx==3.4.2
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.0
PyOpenGL==3.1.7
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
setuptools==75.6.0
six==1.16.0
stable_baselines3==2.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
zipp==3.21.0
